Addition of all scripts for different board elements
Each different scene has its own script to connect it to different classes
Created two superclasses: Placeable for all board elements, and Road for all roads
More functions to be added
Placeable has a member-variable to track which player the element is related to. This value is returned with the method getPlayer, which should be a property, but GoDot has known troubles with that.