Added buildings list
A list to keep track of all built placeables and a method where these will be added to the list and the new nodes/placeables will be created. It is in this method that the main parts to building will be created. The method has commenting indicating roughly what happens where.
Added method where the move will be checked, but that contains no actual checks (always returns true)