Adding in checks reworking build function for building roads
Changed build function and some variables to right case
Changed and added comments
Added road-version of RelevantList
Reworked selection of RelevantList
Added the functions of road instancing
Added function to check whether road can be built
Moved Hexatileboard to be saved as a membervariable in the class. It is given value in the Ready function
To be added is a function to make instances of the roads.