This commit connects the TileClick with BulderNode
The building function works. Some buildings are still very big and need to be scaled down.
There are some changes in TurnManager.gd to get things working properly.
GridClick is also not necessary anymore, so it will be removed.
BuilderNode has a new function to check whether the building and the intended place match, before going into all the other calculations.